Remove some unused dockerfiles
I have not used any of these dockerfiles in quite a while. I no longer
have access to Gurobi (it was through school) and no longer need images
containing the system certificates which was a pretty LLNL specific
problem. The gematria GPU image might be useful but needs to be redone
if I want to use it most likely.